requirements.txt
# Sentient Core v4 - Python Dependencies
# Install with: pip install -r requirements.txt

# Core Speech Recognition & Synthesis
vosk==0.3.45
pvporcupine==3.0.5
pyaudio==0.2.14

# Computer Vision & AI
opencv-python==4.8.1.78
numpy==1.24.3

# System & Hardware Interface
RPi.GPIO==0.7.1
spidev==3.6
smbus2==0.4.3
adafruit-circuitpython-bme280==2.6.19
gpiozero==2.0

# Audio Processing
pydub==0.25.1

# Utilities
pygame==2.5.2
pillow==10.1.0
pyserial==3.5

# Async & Threading
asyncio-mqtt==0.16.2

# ============================================================================
# API INTEGRATION DEPENDENCIES
# ============================================================================

# Environment variable management
python-dotenv==1.0.0

# LLM Backends
ollama==0.3.3
anthropic==0.39.0
openai==1.54.3
groq==0.11.0
google-generativeai==0.8.3

# Web Search & APIs
requests==2.31.0
urllib3==2.1.0

# Database (PostgreSQL)
psycopg2-binary==2.9.9

# Additional utilities for API integration
aiohttp==3.10.5
websockets==12.0
certifi==2024.8.30
